# Conflicts:
#	api/src/main/java/org/api/admin/service/UploadMusicApi.java
#	config/db/mysql/music_box.sql
#	web/src/test/java/org/api/model/LikePlay.java
#	web/src/test/java/org/api/model/lyric/Lyric.java
#	web/src/test/java/org/api/model/song/Al.java
#	web/src/test/java/org/api/model/song/AlbumMeta.java
#	web/src/test/java/org/api/model/song/ArItem.java
#	web/src/test/java/org/api/model/song/ArtistsItem.java
#	web/src/test/java/org/api/model/song/ChargeInfoListItem.java
#	web/src/test/java/org/api/model/song/FreeTrialPrivilege.java
#	web/src/test/java/org/api/model/song/H.java
#	web/src/test/java/org/api/model/song/Hr.java
#	web/src/test/java/org/api/model/song/L.java
#	web/src/test/java/org/api/model/song/M.java
#	web/src/test/java/org/api/model/song/NoCopyrightRcmd.java
#	web/src/test/java/org/api/model/song/OriginSongSimpleData.java
#	web/src/test/java/org/api/model/song/Pc.java
#	web/src/test/java/org/api/model/song/PrivilegesItem.java
#	web/src/test/java/org/api/model/song/SongDetail.java
#	web/src/test/java/org/api/model/song/SongsItem.java
#	web/src/test/java/org/api/model/song/Sq.java
#	web/src/test/java/org/api/utils/RequestMusic163.java

+++ api/src/main/java/org/api/neteasecloudmusic/model/vo/playlistallsong/Hr.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Hr {
     private int br;
-    private int fid;
+    private long fid;
     private int size;
     private int vd;
     private int sr;
