+++ api/src/main/java/org/api/neteasecloudmusic/model/vo/playlistallsong/L.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class L {
     private int br;
-    private int fid;
+    private long fid;
     private Long size;
     private int vd;
     private int sr;
